Tarefa2_2018li1g080

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

posicaoNova

Arguments

:: Jogador
O Jogador que efetuou a Jogada.

-> Direcao
A Direcao na qual o Jogador se pretende mover.

-> Posicao
A nova Posicao do Jogador.

Dados um Jogador e uma Direcao, devolve a Posicao do Jogador movido nessa direção.

temVidas

Arguments

:: Jogador
O Jogador cujo número de vidas queremos verificar.

-> Bool
True se o Jogador tiver vidas, False se não tiver.

Verifica se um dado Jogador ainda tem vidas.

posicaoLivre

Arguments

:: Jogador
O Jogador que efetuou a Jogada.

-> Direcao
A Direcao na qual o Jogador se pretende mover.

-> Estado
O Estado anterior à Jogada.

-> Bool
True se as Pecas à frente do Jogador estiverem Vazias, False se não estiverem.

Verifica se a Posicao para a qual um Jogador se pretende mover se encontra livre de Pecas.

foraDeChoque

Arguments

:: Jogador
O Jogador que efetuou a Jogada.

-> Estado
O Estado anterior à Jogada.

-> Bool
True se o Jogador não estiver dentro de um DisparoChoque de outro Jogador, False se estiver.

Verifica se um dado Jogador se encontra numa zona de DisparoChoque.

semTanques

Arguments

:: Posicao
A Posicao para a qual o Jogador que efetuou a Jogada se pretende mover.

-> Int
O número do Jogador que efetuou a Jogada.

-> [Jogador]
A lista dos Jogadores no Estado atual.

-> Int
O número do Jogador cuja Posicao estamos a comparar com a Posicao para a qual o Jogador que efetuou a Jogada se pretende mover.

-> Bool
True se o Jogador não tiver outros tanques à sua frente, False se tiver.

Verifica se a Posicao para a qual um Jogador se pretende mover se encontra livre de outros Jogadores.