Tarefa3_2018li1g080

Description

Introdução

Nesta tarefa, tivemos que encontrar uma forma de comprimir ao máximo o estado do jogo, por outras palavras, arranjar uma forma de fazer com que o estado ocupasse o mínimo de espaço possível, não perdendo nenhuma da sua informação.

Objetivos

O nosso principal objetivo, para poder atingir uma taxa de compressão máximo, foi conseguir reduzir um estado a um conjunto de dígitos. Estes números poderiam então ser convertidos para números de base superior, ou seja, números que são representados por mais do que 10 caracteres. Por exemplo, no formato binário o número 14 fica 1110, e no formato hexadecimal fica E. Assim, vemos que, quanto mais caracteres diferentes puderem ser usados para representar um número, menos caracteres serão necessários no total. O número máximo de diferentes caracteres a que conseguimos chegar foi 1114701. Antes desta mudança de base, foi necessário reduzir ao máximo o número de dígitos necessários para representar um estado. Para isso, vimo-nos livres de tudo o que não era necessário saber para depois descomprimir o estado. Por exemplo, como a borda do mapa é sempre formada por blocos indestrutíveis, é desnecessáio guardar essa informação. Depois, foi simplesmente uma questão de conseguir extrair o máximo possível de informação a partir de um número o mais reduzido possível de dígitos, o que exigiu alguma criatividade e pensamento fora da caixa.

Discussão e Conclusão

Em retrospetiva, acreditamos que fizemos o melhor que conseguimos com os conhecimentos que possuímos, sendo que até ficamos em primeiro lugar nas classificações quanto à tarefa em questão. Embora, através dos sistemas de avaliação fornecidas pelo professor, nos tenhamos apercebido que a qualidade do código precisava de um pouco mais de trabalho, decidimo-nos focar no nível de compressão. Podemos assim afirmar que consideramos o facto de conseguir equilibrar a compressão/descompressão possível com a qualidade do código como uma tarefa desafiante.

Testes

testesT3 :: [Estado]

Testes unitários da Tarefa 3.

Cada teste é um Estado.

Funções principais da Tarefa 3.

comprime :: Estado -> String

Comprime um Estado para formato textual.

NB: A função show representa um Estado num formato textual facilmente legível mas extenso.

NB: Uma boa solução deve representar o Estado dado no mínimo número de caracteres possível.

descomprime :: String -> Estado

Descomprime um Estado no formato textual utilizado pela função comprime.

NB: A função comprime é válida de for possível recuperar o Estado utilizando a função descomprime, i.e.:

descomprime . comprime = id

NB: Esta propriedade é particularmente válida para a solução pré-definida:

read . show = id

Funções auxiliares da Tarefa 3

Funções auxiliares para a compressão

comprimeMapa :: Mapa -> Int -> String

Comprime um Mapa para formato textual.

comprimeLinha :: [Peca] -> String

Comprime uma linha de um Mapa (uma lista de Peca) para formato textual.

NB: Esta função é usada como função auxiliar da comprimeMapa, para comprimir uma linha do mapa de cada vez.

converteMapa :: String -> String

Converte um Mapa comprimido numa String ainda mais comprimida, usando a base 1114071 para representar o resultado da função compactaMapa.

NB: Desta forma, o nosso Mapa comprimido passa a ser representado por um conjunto de 1114071 diferentes caracteres, em vez de 4.

compactaMapa :: String -> String

Comprime ainda mais o Mapa, após a função comprimeMapa, usando o caracter '3' e outro(s) caracter(es) /= '3' à frente para indicar que uma dada Peca se repete várias vezes seguidas.

Ex: "31012301" é o mesmo que "1111111111112111111"

comprimeJogadores :: [Jogador] -> String

Comprime uma lista de Jogador para formato textual.

NB: Esta função usa a função comprimeJogador para comprimir um Jogador de cada vez.

comprimeJogador :: Jogador -> String

Comprime um Jogador para formato textual.

NB: Os valores auxiliares lenVLS e yxlen servem para a função descomprimeJogador saber se as coordenadas / vidas e munições ocupam um ou dois caracteres.

converteJD :: String -> String

Converte as listas dos Jogadores ou dos Disparos comprimidas, da base 10 para a base 1114071.

comprimeDisparos :: [Disparo] -> String

Comprime uma lista de Disparo para formato textual.

NB: Esta função usa a função comprimeDisparo para comprimir um Disparo de cada vez.

comprimeDisparo :: Disparo -> String

Comprime um Disparo para formato textual.

NB: Esta função usa alguns valores auxiliares, que adiciona à direção, para ao descomprimir poder determinar a direçao, o jogador que disparou, o tipo de disparo ou os ticks apenas com um número. Também é usado um valor auxiliar para verificar ao descomprimir se a posição do disparo ocupa dois ou quatro caracteres (ex: (4,3) vs (12,14)).

Funções auxiliares para a descompressão

desconverteMapa :: String -> String

Converte um Mapa convertido na base 1114071 para uma String de dígitos na base 4, podendo depois ser descomprimido pela função descomprimeMapa.

descompactaMapa :: String -> String

Descomprime as Pecas repetidas representadas pelo caracter '3', para o Mapa poder ser corretamente descomprimido.

descomprimeMapa :: String -> Int -> Mapa

Descomprime um Mapa em formato textual.

NB: Esta função usa a função descomprimeLinha para descomprimir uma linha de cada vez.

descomprimeLinha :: String -> [Peca]

Descomprime uma linha do Mapa (uma lista de Peca) em formato textual.

descomprimeJogadores :: String -> [Jogador]

Descomprime uma lista de Jogadores em formato textual.

NB: Esta função usa a função descomprimeJogador para descomprimir um Jogador de cada vez.

descomprimeJogador :: String -> Jogador

Descomprime um Jogador em formato textual.

desconverteJD :: String -> String

Converte a lista dos Jogadores e dos Disparos comprimidas na base 1114071 para una String equivalente na base 10, para depois poderem ser descomprimidas.

descomprimeDisparos :: String -> [Disparo]

Descomprime uma lista de Disparo em formato textual.

NB: Esta função usa a função descomprimeDisparo para descomprimir um Disparo de cada vez.

descomprimeDisparo :: String -> Disparo

Descomprime um Disparo em formato textual.

Funções auxiliares numéricas

base4PraBase1114071 :: String -> String

Converte um número na base 4 (ex: 1202120) num número na base 1114071 (ex: 5n-F).

NB: Ambos os números têm que estar dentro de uma String.

base4PraBase10 :: String -> Integer

Converte um número na base 4 (ex: 1202120) num número na base 10 (ex: 7323).

NB: O input tem que estar em formato String.

NB: O output é do tipo Integer para evitar as limitações de tamanho do Int.

base10PraBase1114071 :: Integer -> String

Converte um número na base 10 (ex: 7323) num número na base 1114071 (ex: 5n-F).

NB: O output é dado numa String.

base1114071PraBase4 :: String -> String

Converte um número na base 1114071 (ex: 5n-F) num número na base 4 (ex: 1202120).

NB: Ambos os números têm que estar dentro de uma String

base1114071PraBase10 :: String -> Integer

Converte um número na base 1114071 (ex: 5n-F) num número na base 10 (ex: 7323).

NB: O imput tem que estar no formato String.

base10PraBase4 :: Integer -> String

Converte um número na base 10 (ex: 7323) num número na base 4 (ex: 1202120).

NB: O output é dado em formato String.