Tarefa4_2018li1g080

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

mapaTestes :: Mapa

Funções principais da Tarefa 4.

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funções tickChoques, tickCanhoes e tickLasers pela ordem definida.

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

tickLaser :: Estado -> Disparo -> Estado

Função auxiliar da tickLasers, avança apenas um DisparoLaser no tempo.

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

tickCanhao :: Estado -> [Disparo] -> Mapa -> Disparo -> Estado

Função auxiliar da tickCanhoes, avança apenas um DisparoCanhao no tempo.

tickCanhaoMapa :: Mapa -> PosicaoGrelha -> Direcao -> Mapa

Processa os efeitos de um DisparoCanhao no Mapa, ou seja, destroi os Blocos Destrutivelis com os quais o DisparoCanhao colide.

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.

Funções auxiliares da Tarefa 4.

bateNoJogador :: Disparo -> Jogador -> PosicaoGrelha -> Bool

Verifica se um Disparo embate num Jogador

bateEmBlocos :: Disparo -> Mapa -> [PosicaoGrelha]

Devolve os Blocos Destrutivelis que um DisparoLaser destroi.

apagaCanhoes :: Disparo -> [Disparo] -> PosicaoGrelha -> [Disparo]

Destroi os DisparoCanhao que colidem com um DisparoLaser.

inverteDir :: Direcao -> Direcao

Inverte a Direcao dada.

Eg: inverteDir B == C

blocosOndeBate :: PosicaoGrelha -> Direcao -> [Posicao]

Devolve as Posicaoes dos Blocos com os quais um DisparoCanhao com uma dada PosicaoGrelha e Direcao embate.

posJogadoresOndeBate :: PosicaoGrelha -> Direcao -> [PosicaoGrelha]

Devolve as Posicaoes dos Jogadores com os quais um DisparoCanhao com uma dada PosicaoGrelha e Direcao embate.