Tarefa5_2018li1g080

Description

Introdução

A tarefa 5 foi a tarefa na qual criámos o jogo em si, isto é, enquanto que nas outras tarefas definimos o código por detrás do jogo, nesta tarefa tratámos do aspeto visual e do input/output, aquilo que transforma o jogo em algo mais que meras linhas de código.

Objetivos

O nosso principal objetivo era criar um jogo com design intuitivo e minimalista, mas apelativo. Tentámos usar o mínimo de texto possível, e usar imagens que representassem de forma ideal cada objeto, como os tanques, os disparos ou os contadores de vidas e munições. Assim, uma pessoa que nunca tenha jogado o nosso jogo pode aprender a jogar sem ter que ler grandes quantidades de texto, e percebe imediatamente o que cada coisa é ou o que faz.

Escolhemos um tema futurístico para o nosso jogo, sendo a cor principal do jogo o roxo, e demos um efeito neon ao título. Todas as imagens que aparecem no nosso jogo foram criadas por nós ou retiradas da internet, mas tivemos o cuidado de apenas escolher imagens sem direitos de autor, e que pudessem ser reutilizadas.

Queríamos também que o jogador fosse capaz de escolher o mapa onde quer jogar, podendo até criar o seu próprio mapa, se não gostasse de nenhum dos mapas que criámos para o jogo, e que pudesse escolher a cor do seu tanque. Cada tanque tem canhões e lasers com cores próprias, baseadas na cor do próprio tanque. Também demos ao jogador a possibilidade de jogar contra outras pessoas, ou de jogar contra bots, podendo escolher quantos humanos/bots existem em cada partida.

Discussão e conclusão

Gostamos bastante da forma como o nosso jogo ficou, tínhamos uma ideia em mente e acreditamos que fomos capazes de transformar essa ideia num jogo que funciona e é divertido. O haskell não é uma linguagem ideal para criar jogos, mesmo com o gloss, o melhor seria usar algo como o Unity ou o Unreal Engine, ferramentas feitas de raiz para criar jogos, mas mesmo assim este jogo funciona perfeitamente bem, e fomos capazes de fazer tudo o que queríamos.

Em conclusão, esta tarefa foi bastante divertida, já que pudemos dar asas à nossa imaginação e criar algo que outras pessoas pudessem jogar e desfrutar. Foi uma das nossas tarefas preferidas de fazer, precisamente por causa disso. A parte mais difícil foi ter que aprender a usar o gloss, que é um pouco diferente do que tínhamos feito com haskell até ao momento, mas não foi muito difícil apanhar-lhe o jeito, e a partir daí tornou-se fácil de usar.

Documentation

data State

O estado que usámos para guardar todas as variáveis do nosso jogo.

Constructors

State

Fields

estadoState :: Estado

O Estado do jogo.

playersBotsList :: [(PlBt, Int, Points)]

A lista dos jogadores / bots que estão a jogar, o seu tanque (índice da sua imagem) e os seus pontos.

currentMenu :: Menu

O Menu atual.

isMoving :: [(Int, Direcao, Int)]

Os jogadores que se estão a mover de momento. Isto permite que um clique longo de uma tecla para mover mova o tanque continuamente.

currentTick :: Int

O tick atual.

selectedOption :: Options

A Option atualmente selecionada.

drawnLasers :: [(Disparo, Int)]

Os lasers atualmente no mapa.

cleanMap :: Mapa

O Mapa igual ao do Estado, mas com os blocos destrutíveis intactos, para poder reiniciar o jogo no mesmo mapa.

mapInsts :: [Instrucao]

As instruções necessárias para construir o mapa, caso o mapa seja personalizado, isto é, criado pelo jogador.

data Menu

Os vários menus que o nosso jogo possui.

Constructors

MainMenu

MapMenu

PlayerMenu

GameMenu

GameOver

PlayMenu

MapEditor

CreditsMenu

data Options

As opções que é possível selecionar em cada Menu.

Constructors

VoltarAoMenu

JogarNovamente

Jogar

IniciarJogo

AlterarJogadores

AlterarMapa

Mapa1

Mapa2

Mapa3

Mapa4

MapaP

Player1

Player2

Player3

Player4

Bot2

Bot3

Bot4

OKPlayer

Creditos

SairCreditos

data PlBt

Apenas usado para definir se um jogador é humano ou bot.

Constructors

HumanPlayer

BotPlayer

type Points = Int

Os pontos de um jogador.

mainDisplay :: Display

As definições da janela do jogo.

displayDimension :: (Int, Int)

A dimensão da janela do jogo.

fps :: Int

O framerate do jogo, isto é, quantas vezes a função timeReact é executada num segundo.

playBots :: Estado -> [(PlBt, Int, Points)] -> Int -> Estado

Função que permite aos bots efetuar a sua jogada, usando para isso a função bot da Tarefa6_2018li1g080

addPoints :: [(PlBt, Int, Points)] -> Int -> Estado -> [(PlBt, Int, Points)]

Função que adiciona pontos a cada jogador com base no seguinte sistema:

Bloco destruído: 1 ponto;

Vida retirada a jogador: 4 pontos;

Vidas restantes após fim do jogo: 4 pontos por vida.

addBlockP :: Estado -> Int -> Int

Função auxiliar da função addPoints, apenas adiciona os pontos relativos à destruição de blocos.

addPlayerP :: Estado -> Int -> [Disparo] -> Int

Função auxiliar da função addPoints, apenas adiciona os pontos relativos ao dano a outros jogadores.

colocarJogadores :: [Jogador] -> Mapa -> Int -> [Jogador]

Função que coloca os jogadores nos seus lugares iniciais, que no nosso caso é nos cantos dos mapas para todos os mapas.

resetGame :: Estado -> Mapa -> Estado

Função que permite fazer "reset" do jogo, para quando o jogador pretende jogar novamente.

resetScore :: [(PlBt, Int, Points)] -> [(PlBt, Int, Points)]

Função que faz "reset" da pontuação.

gameOverPoints :: Estado -> [(PlBt, Int, Points)] -> [(PlBt, Int, Points)]

Função que dá os devidos pontos aos jogadores com base nas suas vidas quando o jogo acaba.

isBot :: Int -> [(PlBt, Int, Points)] -> Bool

Função que verifica se um dado jogador (identificável pelo seu número) é um bot ou não.

switchBot :: Int -> [(PlBt, Int, Points)] -> [(PlBt, Int, Points)]

Função que altera o tipo de jogador de bot para humano, e vice-versa.

jogada' :: Int -> Jogada -> Estado -> Estado

Esta função já estava definida na tarefa 2, mas como queríamos que no nosso jogo os disparos de choque durassem mais tempo, sem interferir com o resto, voltámos a defini-la aqui.

Funções principais da tarefa 5

main :: IO ()

Função main, permite que o jogo corra, e permite importar as imagens para o jogo.

inputReact :: Event -> State -> State

Função que executa algo no jogo com base na tecla premida pelo utilizador.

initialState :: State

State inicial do jogo.

drawState :: ([[Picture]], [[Picture]], [[Picture]], [[Picture]]) -> State -> Picture

Função que desenha o jogo na janela do gloss.

timeReact :: Float -> State -> State

Função que permite ao jogo gerir o tempo, isto é, que define a passagem do tempo no jogo.

Funções auxiliares da função drawState

desenhaMapa

Arguments

:: Mapa

-> Int
Nº da linha atual, na recursão

-> Mapa

-> [Picture]

-> [Picture]

Função que desenha o Mapa na janela do gloss.

desenhaLinha :: [Peca] -> (Int, Int) -> Mapa -> [Picture] -> [Picture]

Função que desenha cada linha de um Mapa na janela do gloss.

| NB: Esta função é uma função auxiliar da função desenhaMapa.

sc :: Mapa -> Picture -> Picture

Função que permite redimensionar uma Picture para a escala do jogo.

gameScale :: Mapa -> Float

A escala do jogo, com base no tamanho do mapa (quanto maior o mapa, menor a escala).

pieceSide :: Mapa -> Float

O tamanho que cada Peca do Mapa deve ter na janela do gloss.

desenhaJogadores

Arguments

:: [Jogador]
A lista dos jogadores.

-> Int
O nº do jogador que será desenhado na função desenhaJogador.

-> Mapa
O mapa do jogo.

-> [(PlBt, Int, Points)]
A lista que contém o tanque escolhido por cada jogador.

-> [[Picture]]
As imagens relativas aos jogadores.

-> [Picture]
Os jogadores e as suas informações (vidas,munições,etc) para serem desenhados na janela do gloss.

Função que desenha os Jogadores na janela do gloss.

desenhaJogador :: Jogador -> Int -> Mapa -> [(PlBt, Int, Points)] -> [[Picture]] -> Picture

Função que desenha um Jogador na janela do gloss.

NB: Esta função é uma função auxiliar da função desenhaJogadores.

rotatePic :: Direcao -> Picture -> Picture

Função que permite rodar uma Picture, assumindo que a Direcao original é C

desenhaPontos :: [(PlBt, Int, Points)] -> [Picture]

Função que desenha os pontos de cada Jogador na janela do gloss.

desenhaDisparos :: [Disparo] -> [Jogador] -> [(PlBt, Int, Points)] -> Mapa -> [[Picture]] -> [Picture]

Função que desenha os Disparos na janela do gloss.

desenhaDisparo :: Disparo -> [Jogador] -> [(PlBt, Int, Points)] -> Mapa -> [[Picture]] -> Picture

Função qu desenha um Disparo na janela do gloss.

NB: Esta função é uma função auxiliar da função desenhaDisparos.

desenhaEditor :: Editor -> [Picture] -> [Picture]

Função que desenha o Editor de Mapas na janela do gloss.

mapas :: [Mapa]

Os mapas que existem no nosso jogo.

mapa1 :: Mapa

mapa2 :: Mapa

mapa3 :: Mapa

mapa4 :: Mapa

mapa5 :: Mapa

mapa6 :: Mapa

Ignorar estas funções, apenas servem para abreviar coisas mais compridas.

fi :: Int -> Float

t5r :: String