Tarefa6_2018li1g080

Description

Introdução

Esta tarefa consistiu na criação de um bot, isto é, de um "jogador" capaz de jogar sem qualquer input humano, uma espécie de inteligência artificial primitiva. Esta tarefa foi talvez a mais complicada, no sentido em que nos obrigou a pensar em muitas variáveis e possibilidades para o comportamento do bot.

Objetivos

Para nós, o objetivo foi criar um bot que conseguisse fazer tudo o que um jogador humano consegue, que fosse o melhor possível, mas que conseguisse ser derrotado por um humano, senão o jogo ia passar a ser frustrante, em vez de divertido. Para cumprir esse objetivo, definimos várias funções para o bot saber o que fazer. Por exemplo, se existir um disparo a dirigir-se para o bot, este tenta desviar-se. Se não houver nenhum perigo iminente, este tenta encontrar outros jogadores ou blocos destrutíveis, para tentar ganhar pontos. A ordem pela qual estas funções aparecem na função bot define as prioridades do mesmo. No nosso caso, as prioridades do nosso bot são, mais ou menos por esta ordem:

Desviar-se de disparos iminentes;

Disparar em direção a um jogador que se encontre perto do bot;

Disparar em direção a blocos destrutíveis que se encontrem perto do bot;

Verificar se existem, e, caso existam, mover-se para um local onde existam blocos destrutíveis perto;

Mover-se para perto de outros jogadores, para os poder matar.

Se o bot não conseguir fazer nenhuma destas coisas, move-se pelo mapa sem um destino concreto, mas na grande maioria dos casos ele consegue sempre fazer pelo menos uma coisa.

Discussão e conclusão

Infelizmente, tivemos que fazer com que o bot apenas detetasse blocos destrutíveis a um raio de 8 blocos dele, já que um número mais elevado fazia com que o jogo ficasse muito lento, principalmente em mapas grandes. Esta limitação impede que o nosso bot tenha o potencial máximo que podia ter, mas da forma que está já consegue ganhar contra jogadores humanos, em quase todas as partidas, e já ficou em primeiro lugar no ranking da tarefa 6.

Em conclusão, apesar de não termos conseguido fazer tudo o que pretendíamos, devido sobretudo às limitações do haskell, fomos capazes de cumprir o nosso objetivo, e estamos orgulhosas da forma como o nosso bot ficou. Esta tarefa foi um grande desafio, mas no bom sentido, já que nos permitiu testar as nossas capacidades de formas que as outras tarefas não o fizeram, e obrigou-nos a pensar um pouco na forma como a inteligência artificial funciona, e a valorizar bastante os avanços nessa área.

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Define um robot capaz de jogar autonomamente o jogo.

jogadorAdjacente :: PosicaoGrelha -> [Jogador] -> Bool

Verifica se o bot tem algum Jogador encostado a ele, mas que não se encontra diretamente de um dos seus lados.

distMan :: (Int, Int) -> (Int, Int) -> Int

Calcula a distância de Manhattan entre duas Posicaoes.

dispararAdjacente :: PosicaoGrelha -> Direcao -> [Jogador] -> Maybe Jogada

Dispara contra um Jogador que se encontre adjacente ao bot.

canhaoPerto :: PosicaoGrelha -> Int -> Mapa -> [Disparo] -> Bool

Verifica se existe um disparo de canhão a menos de 4 blocos do bot, que não seja do próprio bot, e a mover-se na direção do bot.

desviar :: PosicaoGrelha -> Direcao -> Int -> Estado -> Jogada

Tenta desviar o bot de um disparo de canhão iminente, ou disparar outro canhão para destruir o que se aproxima.

muitosCanhoesAFrente :: PosicaoGrelha -> Direcao -> Estado -> Bool

Verifica se existem vários canhões a dirigir-se em direção ao bot.

canhaoAFrente :: PosicaoGrelha -> Direcao -> PosicaoGrelha -> Direcao -> Bool

Verifica se existe um DisparoCanhao à frente do bot.

canhaoVemDe :: Direcao -> PosicaoGrelha -> Direcao -> PosicaoGrelha -> Bool

Dada uma Direcao, a função retorna True se existir um DisparoCanhao a dirigir-se para o bot vindo dessa direção.

desviarDe :: (Direcao, PosicaoGrelha) -> (PosicaoGrelha, Direcao) -> Estado -> Jogada

Tenta desviar o bot de um DisparoCanhao iminente.

noCanto :: PosicaoGrelha -> PosicaoGrelha -> Bool

Devolve True caso exista um DisparoCanhao a tocar num dos cantos do tanque do bot.

sairDoCanto :: PosicaoGrelha -> PosicaoGrelha -> Direcao -> Jogada

Desvia o bot de um DisparoCanhao que esteja a tocar num dos seus cantos.

livrePara :: Direcao -> PosicaoGrelha -> Estado -> Bool

Verifica se uma dada Direcao se encontra livre para o bot nela se deslocar.

semilivrePara :: Direcao -> PosicaoGrelha -> Estado -> Bool

Semelhante à função livrePara, mas apenas verifica os dois blocos mais próximos do bot, em vez de quatro.

semDisparosPara :: Direcao -> PosicaoGrelha -> [Disparo] -> Bool

Verifica se existe algum DisparoCanhao numa dada Direcao, se não existir nenhum devolve True.

vaziasPara :: Direcao -> PosicaoGrelha -> Mapa -> Bool

Verifica se existem Peca não Vazias nos 4 blocos adjacentes ao bot numa dada Direcao.

semivaziasPara :: Direcao -> PosicaoGrelha -> Mapa -> Bool

Verifica se existem Peca não Vazias nos 2 blocos adjacentes ao bot numa dada Direcao.

viradoParaCanhao :: PosicaoGrelha -> Direcao -> Estado -> Bool

Verifica se o bot está virado para um DisparoCanhao.

virarParaCanhao :: PosicaoGrelha -> Estado -> Jogada

Vira o bot na Direcao de um DisparoCanhao.

jogadorPertoChoque :: PosicaoGrelha -> Mapa -> [Jogador] -> Bool

Verifica se existe algum Jogador perto o suficiente do bot para ser afetado pelo DisparoChoque.

jogadorAVolta :: PosicaoGrelha -> Estado -> Bool

Verifica se existe um Jogador contra o qual o bot pode disparar imediatamente (ou após virar-se).

semChoque :: Int -> [Disparo] -> Bool

Verifica se o bot tem algum DisparoChoque disparado por si atualmente no mapa.

vaziasAte :: PosicaoGrelha -> [Posicao] -> Mapa -> Bool

Verifica se o caminho entre uma PosicaoGrelha e outras Posicaoes se encontra livre.

posJogadoresAVolta :: PosicaoGrelha -> [Jogador] -> [PosicaoGrelha]

Devolve as posicões dos Jogadores à volta do bot.

disparar :: PosicaoGrelha -> Direcao -> Estado -> Maybe Jogada

Dispara um DisparoCanhao contra um Jogador.

jogadorAFrente :: PosicaoGrelha -> Direcao -> [Jogador] -> Mapa -> Bool

Verifica se existe um Jogador à frente do bot.

virarParaJogador :: PosicaoGrelha -> Direcao -> [Jogador] -> Mapa -> String

Devolve a Direcao para a qual o bot se deve virar para estar de frente a um Jogador, em forma de String.

NB:Caso não haja jogadores para os quais o bot se possa virar, a função devolve X, que será interpretado pela função disparar como Nothing.

destrutiveisAVolta :: PosicaoGrelha -> Estado -> Bool

Verifica se existem Blocos Destrutivelis que o bot possa destruir ao seu redor.

bateEmBlocosX :: PosicaoGrelha -> Direcao -> Mapa -> [PosicaoGrelha]

Igual à função bateEmBlocos da Tarefa4_2018li1g080, mas esta recebe uma PosicaoGrelha e uma Direcao em vez de um Disparo. Devolve as Posicaoes dos Blocos Destrutivelis nessa Direcao.

destruir :: PosicaoGrelha -> Direcao -> Int -> Estado -> Jogada

Destroi o(s) Bloco(s) Destrutivel(eis) à volta do bot.

destrutiveisNoMapa :: PosicaoGrelha -> Estado -> Bool

Verifica se existem Blocos Destrutivelis no mapa.

moverPraDestrutivel :: PosicaoGrelha -> Direcao -> Int -> Estado -> Maybe Jogada

Movimenta o bot para onde possa destruir Blocos Destrutivelis.

posicoesOtimas :: PosicaoGrelha -> Mapa -> [Disparo] -> [PosicaoGrelha]

Devolve as PosicaoGrelhas de onde o bot consegue destruir Blocos.

moverPraJogador :: PosicaoGrelha -> Direcao -> Int -> Estado -> Maybe Jogada

Movimenta o bot para uma PosicaoGrelha onde esteja outro Jogador.

posicoesJog :: [Jogador] -> PosicaoGrelha -> Int -> Int -> [PosicaoGrelha]

Devolve as PosicaoGrelhas dos Jogadores que ainda estão vivos e que não são o próprio bot.

moverPara :: PosicaoGrelha -> Direcao -> PosicaoGrelha -> Mapa -> Maybe Jogada

Move o bot de uma PosicaoGrelha para outra PosicaoGrelha, se encontrar um caminho possível.

diretoPara :: Direcao -> PosicaoGrelha -> [PosicaoGrelha] -> Mapa -> Bool

Verifica se existe um caminho direto entre duas PosicaoGrelhas.

indiretoPara :: Direcao -> PosicaoGrelha -> [PosicaoGrelha] -> Mapa -> Bool

Verifica se existe um caminho indireto (ou seja, um caminho que envolve mudar de Direcao) entre duas PosicaoGrelhas.

posAVolta :: PosicaoGrelha -> Mapa -> [Posicao]

Devolve todas as PosicaoGrelhas à volta de uma PosicaoGrelha que se encontrem livres.

posAcima :: PosicaoGrelha -> Mapa -> [Posicao]

Devolve todas as PosicaoGrelhas acima de uma PosicaoGrelha que se encontrem livres.

posAbaixo :: PosicaoGrelha -> Mapa -> [Posicao]

Devolve todas as PosicaoGrelhas abaixo de uma PosicaoGrelha que se encontrem livres.

posEsq :: PosicaoGrelha -> Mapa -> [Posicao]

Devolve todas as PosicaoGrelhas à esquerda de uma PosicaoGrelha que se encontrem livres.

posDir :: PosicaoGrelha -> Mapa -> [Posicao]

Devolve todas as PosicaoGrelhas à direita de uma PosicaoGrelha que se encontrem livres.

deambular :: PosicaoGrelha -> Direcao -> Estado -> Maybe Jogada

Faz com que o bot ande pelo Mapa aleatoriamente, sem destino definido. Esta função só é executada se o bot não se conseguir deslocar para nenhum Jogador / 'Bloco Destrutivel'.